[Style] Rename files
