+++ src/components/blocks/GPTLoadingBlock.tsx
