+++ src/components/blocks/MessageBlock.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import styled from "styled-components";
-import ChoiceButton from "./choicebutton";
+import ChoiceButton from "./ChoiceButtonBlocl";
 import questions from "@assets/data/questions.json";
 import SystemThumb from "@assets/images/SystemThumb.png";
-import WidthBox from "./widthblock";
+import WidthBox from "./WidthBlock";
 import { useRecoilState } from "recoil";
 import { depthState, roleState, toState } from "atoms/messagestates";
