+++ src/components/blocks/UserChoiceButtonBlock.tsx
