+++ src/styles/AlertModalStyle.ts
