+++ src/styles/CopyModalStyle.ts
