+++ src/styles/HeadersStyle.ts
